created individual components for dailyGame and practiceGame
