+++ src/App.css
@@ -38,24 +38,19 @@
 .top-section {
   display: flex;
   flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
+  justify-content: center;
   text-align: left;
+  place-items: center;
   gap: 1vw;
 }
 
 .page-center {
   display: flex;
   flex-direction: column;
-  align-items: center;
-  justify-content: center;
+  place-items: center;
   text-align: center;
 }
 
 ul {
   list-style-type: none;
-}
-
-.card {
-  padding: 2em;
 }
